+++ src/app/components/jobIntroduction.tsx
@@ -10,13 +10,12 @@ const JobIntroduction = ({ image, title, text }: Props) => {
 	return (
 		<div className="mx-[8%] my-[8%] flex flex-col rounded-md bg-white p-4">
 			{/* 画像の表示 */}
-			<div className="mb-4 w-full">
+			<div className="relative mb-4 aspect-[16/9] w-full">
 				<Image
 					src={image}
 					alt={title}
-					width={640}
-					height={360}
-					className="rounded-md"
+					fill
+					className="rounded-md object-cover"
 				/>
 			</div>
 			{/* テキストエリア */}
